+++ lib/src/dpop/dpop_token_generator.dart
@@ -31,11 +31,11 @@ import 'dart:convert';
 
 import 'package:crypto/crypto.dart';
 import 'package:dart_jsonwebtoken/dart_jsonwebtoken.dart';
-import 'package:fast_rsa/fast_rsa.dart';
 import 'package:logging/logging.dart';
 import 'package:uuid/uuid.dart';
 
 import 'package:solid_auth/src/dpop/dpop_key_manager.dart';
+import 'package:solid_auth/src/dpop/rsa_key_utils.dart';
 
 final _log = Logger('solid_auth.DpopTokenGenerator');
 const _uuid = Uuid();